st/stcompare/requirements.txt
akshare>=1.17.0
apscheduler>=3.10.4
pandas>=2.2.0
PyYAML>=6.0.1
rich>=13.7.0
pytest>=8.2.0
